Set up grid in webpage
First webpage was set up. It has a grid with three columns, name, status and contact. Each take a Resource object and take each corresponding method. The AvailaDatabase class was created to load, create and enter objects to and from the database. The methods are currently only a skeleton, the Database needs to be setup before anything else can happen. Finally, the Resource class got a Status and Contact field added. This is because when calling for Status and Contact from the grid the compiler throws an exception when trying to call a method from a subclass of Resource.