Getting off for night
Getting off for the night. Created the basic Availaboard class add flushed out the first method of the AvailaboardSQLConnection class. After creating a css file for the grid and getting it working, I will move onto working on a DataBase locally. This will be temporary until there is a VM set up where the application can use it in the cloud. The VM will probably be Azure because Hillel may get it for free.