Styled and added functionality to grid
Added and finished the skeleton for the available and unavailable status for the grid. Now trying to create a button that allows a user to click on a name and route them to a page with the information about that person. So for I can not find something that allows you to create dynamic pages at runtime which means setting the parameters of a user to a page would be impossible.