Updated ResourcePage + Started database
Resource page now get's a Resource ID on initialization. Two database tables have been set up, User and Resource. The resource column has Contact information, Name and a primary key called ResourceID. The user table uses the ResourceID as a foreign key and has information about the User such as firstname, lastname, email ect. The database has started to be setup, it uses a config file to get the database username password and URL. That config file has a class to get the information about it. The log class was also created and will have ability's to log actions that are happening in colors to make tracking events easy on the console.