Fixed database + started generic programming
Finally fixed the database! The problem was that I had quotes around the username, password and URL because the program was getting them from a config file and automatically added quotes around them. Today I also began the generic programming of the project. The goal is to be able to create a class and automatically check it in as a grid object. The reason for this is so I don't have to rewrite each grid object every time a new type of resource get's added. If a new type of resource is created the fields that should be shown just need to be tagged with the annotation of @ResourceInformationLoader. The program is going to have a separate grid for each type of Resource. Finally, the Resource object does not have contact information because that does not make sense.