Began ResourceLoader
For each object being passed into a new grid a feature that is being added is when a user clicks on a row the row expands and shows information about that object type. A way to do it would be to program each field for each class specifically, and although it would have been easier to do it's  bad practice since in the future when adding classes it would be annoying to update it every time. So, instead using the @ResourceLoader annotation I can easily mark the fields that are going to be shown on the click event. I then use some reflection in the ResourceInformationLoader class to get all the fields of that class with the annotations and convert them into textfields, and also get all of there values and add them to the textfields. After I did all of that a "very rare" problem occurred with maven, so I am saving the code before I fix it.