Continued working on createResourceWithID
Tried another way to solve the createResourceWithID problem. Attempted to iterate through each subclass of the Resource class and get each field value. Getting the value is throwing an error so asking arrow on Discord and Stackoverflow to see if I can get the answer. I can always revert to the original method if need be, and modify it with the setAllFields method which, like the name suggests, automatically creates an object and iterates through each field of it and sets each field with data from the database. It's not the optimal solution because each time a class is added on another overloaded addition to that method needs to be added on but it may be necessary. Moral of the story: Java reflection is the WORST. I'll add more documentation in the file about how the methods work later.