Figured out createResourceWithID!
I finally figured out the createMethodWithID! I basically just had to pass in a class and then create an instance of that class and cast it to a Resource object. I then had to iterate through each field in the class and set them. Later I will work on connecting that to the database, but that should be very quick.