Added getResourceGrids + titles to grids
I added a method to the Availaboard class that returns all of the possible grids. It gets all the grids by iterating through all of the subclasses of the Resource class and creating a grid for each of them, then adding it to an ArrayList. A FieldExcludedFromDatabase interface was added so any field that is declared in a subclass of Resource that should not be loaded from that database can be excluded. Finally, titles have been started to be added to the grids. The loadGrid method now requires a class to be passed in so it can set the name of the class as the name of the grid.