Tested getting values from Grid
Added a testing block of code in the ResourceGrid class to attempt to get all the fields and values of the Resource of that grid type when being selected. For some reason the values are all being set to null. I suspect it is because of generic type-erasure (https://docs.oracle.com/javase/tutorial/java/generics/erasure.html) basically generic objects get set to an java.lang.Object causing all of their field values to be lost. That still does not explain why the status and name values successfully get set though.