Finished pop-up windows
Added and finished pop-up windows to the grids. When a user clicks on a name it creates a pop-up windows with all the field names and field values. Each class now has a HashMap of nicknames for the fields so when the fields names are called from the grid it returns a nickname for the fields.