Started permissions for User
Tonight I created a Permission enum and a field in the User class named permissions. The goal is going to be loading permissions from the database so that different types of users have different abilities. I also added a utility method named setType in the AvailaboardSQLConnection that takes in a field, resource and value and sets it. The reason I created it instead of just usiing field.set() is because the JBDC does not support Enums, so this just converts it into an Enum. Currently the values are being set to null so something is wrong.
          Finally, for some reason eclipse is not printing out anything to console. This is super annoying because I can't use print statements to test stuff, and I don't get error stack traces.