Fixed and upgraded to Vaadin 23!
I spent the last four hours fixing the project. I was messing around for a while, screwing with npm (I downloaded a program off github that makes it easy to configure) and the pom.xml. In the end what fixed the project was upgrading it to Vaadin 23 which probably is better long term anyways. That required me to delete the node files and the package-lock.json file and then reinstall it with npm through cmd. After all of that the dark theme variant was not working and after reading through the essay of a stacktrace I figured out that you've got to implement AppShellConfigurator. So even though a lot of files have been updated, added, and deleted, really not that much has happened.