Mostly fixed
I have been messing around for the past 4-5 hours and have been trying to fix the project. (Keep in mind working remotely on a crappy laptop out of DC.) I've updated the project and pom.xml and all is almost well. The only problem is on compilation it throws a ------------------ Frontend compilation failed. ------------------. Very weird. Going to try to figure it out tonight, but I may have to finish it tomorrow.