Fixed outdated node problem
I just published a question on stackoverflow asking about my current bug. But, I did solve the outdated node bug. All I had to do was add a dependency that forced it to update to the right version, instead of letting Vaadin automatically do it. When Vaadin did it the version was outdated. In the meantime I tidied up some code and continued to screw around. The answer to the bug will hopefully be answered in the morning so I can move on with the project.