Styling main layout
The MainLayout class is simply a Parent Layout for all other classes. It consists of a single element; a navbar that pops out of the left side to reveal options for a user to do. Most of this commit is just styling but I also added all of the elements in the navbar to a vertical layout so they get displayed vertically.