Updated CURRENT_USER_SESSION_ATTRIBUTE_KEY to the current session ID and fixed the null User bug
Before the CURRENT_USER_SESSION_ATTRIBUTE_KEY was simply the current users username. The problem with this is a hacker could then stay in the same session indefinitely. Also, a User could stay in that session permanently because a new session ID would never be generated. Instead I used the current request ID as the CURRENT_USER_SESSION_ATTRIBUTE_KEY.
       The other problem was that when a User would try to navigate to the Admin view without being logged in, their CurrentUser User object would never be set meaning when the program tried to access their permissions it would be null and throw a NullPointerException. I fixed it by adding an if statement that created a new User with a User permission if the current user object was null. I may create a new NullUser class in the future.