Fixed User never being null bug
In my last commit, I realized I had a problem with the User object never being null. The reason it was never null is because I had added a if statement to my get() method to return a User object with the permission of User if the CurrentUser session User object was null. This was to avoid the NullPointerException that I would get when seeing if the User was a User or an Admin in my ConfigureUIServiceInitListener class. (When it used to return null, there would be no Permission set which then complained with a NullPointerException. But, later a problem arose when trying to see if

A. A User was logged in
B. A User had a User permission

To fix this I refactored my code by reverting to returning null in my get() method, and simply having a null check in my ConfigureUIServiceInitListener before checking the Permission.