Fixed Navigation Bar not automatically rendering new options when logging in
The Navigation Bar requires a reload to show the specific options for the type of User. The reason for this is because in the MainLayout it overrides the onAttach method. That method is called when a component is added, but since the view is already loaded the method does not get call meaning the Navigation Bar would not get updated. To fix this I just refreshed the page on logging in. Not the most elegant solution, but it works.
     I also applied an Eclipse cleanup which is why so many files have been edited.