Edit button added to Dialog and the Fields are now editable
I spent some time making the fields in the Availability Grid's Dialog editable for Admin Users. It is fully functional, although I will need to add some CSS to make it look nicer. I also am going to add some functionality to it so when a User hits the done button the fields are updated in the database.