Added insertResourceIntoDatabase method
The insertResourceIntoDatabase() now works. I also added a doesUsernameExist() method to check if a Username exists in the database to verify uniqueness. The insertResourceIntoDatatbase method throws an UsernameExistsException if the Username already exists. That exception is caught in the CreateNewAccountView class and it will show an error message when the exception happens.