Replaced all ViewFactory methods with a ViewType factory getter.
Once all of the methods were replaced, I had to update everything implementing them. I've ran into a few hiccups with casting because a lot of the Vaadin methods require the classes to extend from a Component, so I am trying to figure out if there is a bypass so I don't have to cast so much.

I also added a background color to the StatusUsernameContainer in the UserInformationView.