Continued observer pattern
So I fixed the major part of the problem, the Views throwing an exception when being instantiated. I fixed it by putting all my init code inside by init method instead of the constructor.

Now a problem that I have is pushing information around and updating the client. I assume Vaadin has some built-in mechanisms to solve this problem.

The good part is once I have this working all I need to do to have a View be updated is override the Observer interface, and every time I want to update the application call ViewFactory.createViewControllerInstance().notifyObservers