Observer pattern fully works with autmatic notifications
The Views are automatically added to the Subject with the BeforeEnterObserver interface when a User navigates to them, and automatically removes a Observer from a subject with the BeforeLeaveObserver interface. I also changed the Observer interface to extend ViewConfiguration so there would be no conflict of the BeforeEnterObserver implementation in the class.