Adjusted submitButton event documentation and implementation
I separated the blocks of code setting the tempUser fields by the ones setting the User part of it and the ones setting the Resource fields. I also added the User as a CreateNewAccountView Field so it would be instantiated at instantiation time instead of being re-instantiated every time a User clicked on the button.