Fixed the User being able to hit delete when an Admin changes their information
It was a very weird bug. I suspect it's because the ResourceGrid was having a hard time knowing what accessControl to look at so it accidentally thought that the User was an Admin giving them the Admin controls.

To fix it instead of reloading the entire grid I just reloaded the items. This is also more efficient to the webpage.