Added addResourceButton
If the User is an Admin then they can select a Resource type from the drop-down box and press the plus button. It will then bring up a dialog where the User can enter all the information about the Resource.

The ResourceFactory is used to create a Resource with a String.