Fixed non existent Observer being on the Observer list
When a User logged out the View would not be removed from the Observable list. The problem was even though the View was disattached the update method was attempted to be called on it resulting an Error. To fix it every time the logout call was made I just removed the current view from the observable list.