Fixed the Error notification popping up when submitting new information about the User but keeping the same username
The name field of the Resource class was never being set in the UserInformationView, meaning when the Resource was being passed into the updateResourceInDatabase method when the method was checking the CurrentUsers username against the username in the database it would return false causing it to throw an exception.

The simple solution was to set the name field of the Resource.